refactor: split the cards component from eventspanel
