+++ web/src/components/eventslist-component/cards.js
@@ -1,129 +1,32 @@
 import React, {Component} from 'react';
-
+import PropTypes from 'prop-types';
 import {
-    ToggleButton,
-    ToggleButtonGroup,
-    Jumbotron
-} from 'react-bootstrap';
-
-import {
-    Avatar,
+    Avatar, Button,
     Card,
-    CardActionArea,
-    CardMedia,
+    CardActionArea, CardActions,
     CardContent,
-    CardActions,
-    Button,
-    Typography,
-    Grid,
-    List,
-    ListItem,
-    ListItemAvatar,
-    ListItemText
+    CardMedia,
+    Grid, ListItem, ListItemAvatar, ListItemText, Typography,
 } from '@material-ui/core';
 
-import {gql} from 'apollo-boost';
-import { Query } from "react-apollo";
-
-/* Query requests event list to server */
-const EVENT_LIST_REQUEST_QUERY = gql`
-    query{
-        events(pageSize: 9){
-            cursor,
-            hasMore,
-            events{
-                id,
-                title,
-                description
-            }
-        }
-    }`;
 
 class Cards extends Component {
     constructor(props) {
-        // get area selection from props
         super(props);
 
-        // save the arrays of filter names and filter toggle(selected or not) in state
-        const filterNames = ['BookClub', 'Movie', 'Wine', 'SciFi', 'Sport'];
-        const filterToggles = [];
-
-        for (let j = 0; j < filterNames.length; j++)
-            filterToggles.push(false);
-
-        this.state = {
-            filterNames: filterNames,
-            filterToggles: filterToggles,
-        };
     }
 
-    // Function for update filter toggled value to new value
-    toggleHandler = val => {
-        console.log('toggle:', val);
 
-        const newFilterToggles = [];
-        for (let j = 0; j < this.state.filterToggles.length; j++)
-            newFilterToggles.push(false);
-        for (let j = 0; j < val.length; j++)
-            newFilterToggles[val[j]] = true;
-
-        this.setState({
-            filterToggles: newFilterToggles,
-        });
-    };
-
-    // Function for display toggle button group for event filter
-    ToggleButtonGroupControlled = () => {
-        const filtersComp = [];
-
-        // Match each toggle button to filterNames for display
-        for (let i = 0; i < this.state.filterNames.length; i++) {
-            filtersComp.push(
-                <ToggleButton id={i} value={i} key={this.state.filterNames[i]}>
-                    {this.state.filterNames[i]}
-                </ToggleButton>);
-        }
-
-        // Show the filter-button group
-        return (
-            <ToggleButtonGroup type="checkbox" onChange={this.toggleHandler}>
-                {filtersComp}
-            </ToggleButtonGroup>);
-    };
-
-    // Not Used:: Featured Information component for landing page
-    FeaturedInfoComponent = () => {
-        let featuredInfo = [];
-        featuredInfo.push(
-        )
-        return featuredInfo;
-    };
-
-    // Will add Skeleton component during loading
-
-    Items = () => {
-        return (<Query query={EVENT_LIST_REQUEST_QUERY}>
-            {({loading, error, data}) => {
-                if (loading) return "Loading...";
-                if (error) return `Error! ${error.message}`;
-
-
-                return this.CardsComponent(data.events.events);
-            }}
-        </Query>);
-    };
-
-    //will add Skeleton component during loading
-    CardsComponent = (items) => {
+    render() {
         let cards = [[]];
 
-        for (let i = 0; i < items.length; i++) {
+        for (let i = 0; i < this.props.children.length; i++) {
             if (i % 3 === 0)
                 cards.push([]);
 
             // Push each card component in cards
             cards[cards.length - 1].push(
-                <Grid item xs={4} key={items[i].id}>
+                <Grid item xs={4} key={this.props.children[i].id}>
                     <Card style={{marginBottom: '10px'}}>
                         <CardActionArea>
                             {/* Image section of Card */}
@@ -140,10 +43,10 @@ class Cards extends Component {
                                     <ListItemAvatar>
                                         <Avatar alt="Example User Name" src={require('../images/' + (i + 1) + '.jpg')} />
                                     </ListItemAvatar>
-                                    <ListItemText primary={items[i].title} secondary="July 20, 2014" />
+                                    <ListItemText primary={this.props.children[i].title} secondary="July 20, 2014" />
                                 </ListItem>
                                 <Typography style={{height: '150px'}}>
-                                    {items[i].description}
+                                    {this.props.children[i].description}
                                 </Typography>
                             </CardContent>
                         </CardActionArea>
@@ -169,22 +72,9 @@ class Cards extends Component {
         }
 
         return decks;
-    };
-
-    // Render of cards component
-    render() {
-        return (
-            <div>
-                <p>This is {this.props.area}</p>
-                <this.ToggleButtonGroupControlled/>
-                <br/>
-                <hr/>
-                {this.Items()}
-            </div>
-        );
     }
 }
 
-Card.propTypes = {};
+Cards.propTypes = {};
 
 export default Cards;
