+++ web/src/pages/eventslist.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 
-import Cards from "../components/eventslist-component/cards";
+import Eventspanel from "../components/eventslist-component/eventspanel";
 import Paper from "@material-ui/core/Paper";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
@@ -69,7 +69,7 @@ class Eventslist extends Component {
     // match the aligned cards to each place panel
     for(let j = 0; j < this.state.places.length; j++){
       components.push(<TabPanel value={this.state.selectedTap} index={j} key={this.state.places[j].eventKey}>
-        <Cards area={this.state.places[j].eventKey} />
+        <Eventspanel area={this.state.places[j].eventKey} />
       </TabPanel>);
     }
